# docs/openapi/catalogo.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Catálogo de filmes — API pública",
    "version": "1.0.0",
    "description": "Sessão pelo cookie HttpOnly token. Faça login em /login.html no mesmo navegador ou execute POST /api/auth/login com uma conta local. Não cole o cookie no Authorize. GET /api/auth/me sem sessão retorna 401 e é uma chamada segura para demonstrar Try it out. Papel comum: usuario; administrador: admin. Alterações de dados devem ser testadas somente no ambiente local."
  },
  "servers": [
    {
      "url": "/",
      "description": "Mesma origem do catálogo (local ou publicada)"
    }
  ],
  "components": {
    "schemas": {
      "Mensagem": {
        "type": "object",
        "properties": {
          "mensagem": {
            "type": "string"
          }
        },
        "required": [
          "mensagem"
        ]
      },
...
      "Comentario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "example": 1
          },
          "tmdb_movie_id": {
            "type": "integer",
            "minimum": 1,
            "example": 1
          },
          "texto": {
            "type": "string",
            "example": "Comentário de exemplo"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tmdb_movie_id",
          "texto"
        ]
      },
...
    },
    "securitySchemes": {
      "sessao": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "Cookie HttpOnly criado por POST /api/auth/login. Faça login pelo formulário /login.html no mesmo navegador ou execute o login no Swagger. O navegador envia o cookie automaticamente; não leia nem cole o cookie no Authorize. SameSite=Lax; Secure em produção; duração 8 horas."
      }
    }
  },
  "paths": {
...
    "/api/comentarios": {
      "get": {
        "tags": [
          "Comentários"
        ],
        "summary": "Listar comentários próprios",
        "description": "Somente comentários do usuário autenticado; mesmo admin não lista comentários de outras contas. Ordem criado_em DESC. Requer sessão autenticada (cookie HttpOnly token).",
        "operationId": "get_api_comentarios",
        "security": [
          {
            "sessao": []
          }
        ],
        "responses": {
          "200": {
            "description": "Comentários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comentarios": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Comentario"
                      }
                    }
                  },
                  "required": [
                    "comentarios"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Usuário não autenticado. / Token não informado. / Sessão inválida ou expirada. / Token com usuário inválido. / Usuário não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Usuário não autenticado."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Token não informado."
                    }
                  },
                  "2": {
                    "value": {
                      "mensagem": "Sessão inválida ou expirada."
                    }
                  },
                  "3": {
                    "value": {
                      "mensagem": "Token com usuário inválido."
                    }
                  },
                  "4": {
                    "value": {
                      "mensagem": "Usuário não encontrado."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao buscar comentários. / Erro interno do servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Erro ao buscar comentários."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Erro interno do servidor."
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Serviço de autenticação indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Serviço de autenticação indisponível."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/comentarios/{id}": {
      "delete": {
        "tags": [
          "Comentários"
        ],
        "summary": "Excluir comentário",
        "description": "Somente autor ou admin. A permissão é conferida no backend; negação gera ACAO_NEGADA. Requer sessão autenticada (cookie HttpOnly token).",
        "operationId": "delete_api_comentarios_id",
        "security": [
          {
            "sessao": []
          }
        ],
        "responses": {
          "200": {
            "description": "Comentário removido com sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "mensagem": "Comentário removido com sucesso!"
                }
              }
            }
          },
          "400": {
            "description": "ID do comentário inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "ID do comentário inválido."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Usuário não autenticado. / Token não informado. / Sessão inválida ou expirada. / Token com usuário inválido. / Usuário não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Usuário não autenticado."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Token não informado."
                    }
                  },
                  "2": {
                    "value": {
                      "mensagem": "Sessão inválida ou expirada."
                    }
                  },
                  "3": {
                    "value": {
                      "mensagem": "Token com usuário inválido."
                    }
                  },
                  "4": {
                    "value": {
                      "mensagem": "Usuário não encontrado."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Você não tem permissão para excluir este comentário.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Você não tem permissão para excluir este comentário."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Comentário não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Comentário não encontrado."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao excluir comentário. / Erro interno do servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Erro ao excluir comentário."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Erro interno do servidor."
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Serviço de autenticação indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Serviço de autenticação indisponível."
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Inteiro positivo.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          }
        ]
      }
    },
...
  },
...
}